terraform/shared/apt-repo-indexer/s3apt/requirements.txt
boto3==1.9.174
botocore==1.12.174
chardet==3.0.4
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
python-debian==0.1.35
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
